refactor: Extract runSnapshotScan, runIdleScan to snapshot.go
main.go: 1256 → 1210 lines
